Update Control.hs
If the incoming OSC value is integer format, and Tidal is converting it to a Pattern Double, then it's likely a MIDI value >float conversion.  So dividing by 127 here is more convenient for the end user - with this change both `cI` and `cF` will produce reasonable results from incoming Integer messages.